target: Remove last accesses to private structure from outside.
Only include target_internal.h from inside target files.
