+++ src/cortexm.c
@@ -29,10 +29,9 @@
 #include "exception.h"
 #include "adiv5.h"
 #include "target.h"
-#include "command.h"
+#include "target_internal.h"
 #include "gdb_packet.h"
 #include "cortexm.h"
-#include "morse.h"
 
 #include <unistd.h>
 
@@ -507,7 +506,6 @@ static int cortexm_halt_wait(target *t)
 	case EXCEPTION_ERROR:
 		/* Oh crap, there's no recovery from this... */
 		target_list_free();
-		morse("TARGET LOST.", 1);
 		return SIGLOST;
 	case EXCEPTION_TIMEOUT:
 		/* Timeout isn't a problem, target could be in WFI */
