+++ src/lpc15xx.c
@@ -21,6 +21,7 @@
 
 #include "general.h"
 #include "target.h"
+#include "target_internal.h"
 #include "cortexm.h"
 #include "lpc_common.h"
