+++ src/lpc_common.c
@@ -18,6 +18,7 @@
  */
 #include "general.h"
 #include "target.h"
+#include "target_internal.h"
 #include "cortexm.h"
 #include "lpc_common.h"
