+++ src/samd.c
@@ -33,10 +33,8 @@
  */
 
 #include "general.h"
-#include "jtagtap.h"
-#include "adiv5.h"
 #include "target.h"
-#include "command.h"
+#include "target_internal.h"
 #include "gdb_packet.h"
 #include "cortexm.h"
